Refactor tool name
Modify class, package, method and related names based on the tool's
name change. Minor change to fix code lost in previus commit.
